Refactor SocialLoginMapper
SocialLoginMapper -> separate ABC and Interface to their own classes

Closes #98
